+++ src/bin/rbw/commands.rs
@@ -425,6 +425,7 @@ pub fn config_set(key: &str, value: &str) -> anyhow::Result<()> {
                 config.lock_timeout = timeout;
             }
         }
+        "pinentry" => config.pinentry = value.to_string(),
         _ => return Err(anyhow::anyhow!("invalid config key: {}", key)),
     }
     config.save()?;
@@ -449,6 +450,7 @@ pub fn config_unset(key: &str) -> anyhow::Result<()> {
         "lock_timeout" => {
             config.lock_timeout = rbw::config::default_lock_timeout()
         }
+        "pinentry" => config.pinentry = rbw::config::default_pinentry(),
         _ => return Err(anyhow::anyhow!("invalid config key: {}", key)),
     }
     config.save()?;